Update research hub page
